+++ lib/mongo_client.js
@@ -10,7 +10,8 @@ var parse = require('./url_parser')
   , MongoError = require('mongodb-core').MongoError
   , Db = require('./db')
   , f = require('util').format
-  , shallowClone = require('./utils').shallowClone;
+  , shallowClone = require('./utils').shallowClone
+  , authenticate = require('./authenticate');
 
 /**
  * @fileOverview The **MongoClient** class is a class that allows for making Connections to MongoDB.
@@ -359,7 +360,7 @@ function connectHandler(options, callback) {
     }
 
     // Authenticate
-    authentication_db.authenticate(options.user, options.password, options, function(err, success){
+    authenticate(authentication_db, options.user, options.password, options, function(err, success) {
       if(success){
         process.nextTick(function() {
           try {
