+++ package.json
@@ -15,6 +15,7 @@
   "dependencies": {
     "es6-promise": "3.2.1",
     "mongodb-core": "2.1.10",
+    "mongodb-core": "christkv/mongodb-core#2.0",
     "readable-stream": "2.2.7"
   },
   "devDependencies": {
